Fix layout rendering of Visualization #2 on mobile
It appears that the specification of viewBox in the svg specification
to render the navigation map properly was causing issues with the
rendering of the bar graph.

Removed it and positioned the navigation map in a more artifical manner
using proportional offsets based on the width of the container being
rendered on the device.